refactor: #13-redis 폴더 구조 변경 반영
